requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
Faker==28.0.0
gprof2dot==2024.6.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
openai==1.43.0
psycopg2==2.9.9
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
